pipx_info: factored process_list out
